refactor: reorganizing solution and projects
