--- TaskManager.ConsoleInteraction/Title.cs
@@ -1,18 +1,0 @@
-﻿
-
-namespace TaskManager.ConsoleInteraction
-{
-    public class Title
-    {
-        //https://www.asciiart.eu/text-to-ascii-art
-        public static void AllUsers()
-        {
-            Console.WriteLine(" _   _             __       _           \r\n| | | |___ _   _  /_/_ _ __(_) ___  ___ \r\n| | | / __| | | |/ _` | '__| |/ _ \\/ __|\r\n| |_| \\__ \\ |_| | (_| | |  | | (_) \\__ \\\r\n \\___/|___/\\__,_|\\__,_|_|  |_|\\___/|___/");
-        }
-
-        public static string MainMenu()
-        {
-            return " __  __       _         __  __                  \r\n|  \\/  | __ _(_)_ __   |  \\/  | ___ _ __  _   _ \r\n| |\\/| |/ _` | | '_ \\  | |\\/| |/ _ \\ '_ \\| | | |\r\n| |  | | (_| | | | | | | |  | |  __/ | | | |_| |\r\n|_|  |_|\\__,_|_|_| |_| |_|  |_|\\___|_| |_|\\__,_|";
-        }
-    }
-}
