+++ TaskManager.DomainLayer/Model/Developer.cs
@@ -1,16 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using TaskManager.ConsoleInteraction.Components;
 
-namespace TaskManager.Core.People
+namespace TaskManager.DomainLayer.Model
 {
     internal class Developer : User
     {
         public Developer(string newName, string newLogin, string newPassword, string? newEmail = null) : base(newName, newLogin, newPassword, newEmail)
         {
             SetJob(JobEnum.Developer);
+        }
+
+        public override void Greeting()
+        {
+            Console.WriteLine(Title.GreetingDev());
         }
     }
 }
