+++ TaskManager.DomainLayer/Model/IUser.cs
@@ -1,7 +1,8 @@
-﻿namespace TaskManager.Core.People
+﻿namespace TaskManager.DomainLayer.Model
 {
     internal interface IUser
     {
         bool IsValidEmail(string? value);
+        void Greeting();
     }
 }
