+++ TaskManager.DomainLayer/Model/JobEnum.cs
@@ -1,0 +1,9 @@
+﻿
+namespace TaskManager.DomainLayer.Model
+{
+    internal enum JobEnum
+    {
+        Developer = 1,
+        TechLeader = 2
+    }
+}
