+++ TaskManager.DomainLayer/Model/TechLeader.cs
@@ -1,16 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿
+using TaskManager.ConsoleInteraction.Components;
 
-namespace TaskManager.Core.People
+namespace TaskManager.DomainLayer.Model
 {
     internal class TechLeader : User
     {
         public TechLeader(string newName, string newLogin, string newPassword, string? newEmail = null) : base(newName, newLogin, newPassword, newEmail)
         {
             SetJob(JobEnum.TechLeader);
+        }
+
+        public override void Greeting()
+        {
+            Console.WriteLine(Title.GreetingTechLead());
         }
     }
 }
