+++ TaskManager.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34316.72
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskManager.Core", "TaskManager.Core\TaskManager.Core.csproj", "{CB0964B4-6B0D-460B-A8B2-B4ED0CFB76DB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskManager.DomainLayer", "TaskManager.DomainLayer\TaskManager.DomainLayer.csproj", "{CB0964B4-6B0D-460B-A8B2-B4ED0CFB76DB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskManager.ConsoleInteraction", "TaskManager.ConsoleInteraction\TaskManager.ConsoleInteraction.csproj", "{8D4F01F8-BE88-4AA3-9FC4-822D2443CC79}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaskManager.UI", "TaskManager.UI\TaskManager.UI.csproj", "{F4377612-2473-46E7-B19A-CE9FEB46502B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskManager.Infrastructure", "TaskManager.Infrastructure\TaskManager.Infrastructure.csproj", "{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaskManager.Infrastructure", "TaskManager.Infrastructure\TaskManager.Infrastructure.csproj", "{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaskManager.UI", "TaskManager.UI\TaskManager.UI.csproj", "{FE9BE527-988E-4129-AA50-22AFC1C26918}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,14 +25,14 @@ Global
 		{8D4F01F8-BE88-4AA3-9FC4-822D2443CC79}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D4F01F8-BE88-4AA3-9FC4-822D2443CC79}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D4F01F8-BE88-4AA3-9FC4-822D2443CC79}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F4377612-2473-46E7-B19A-CE9FEB46502B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4377612-2473-46E7-B19A-CE9FEB46502B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4377612-2473-46E7-B19A-CE9FEB46502B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4377612-2473-46E7-B19A-CE9FEB46502B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5BAD7AF7-C49E-4609-9960-9E6D6191F0A3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FE9BE527-988E-4129-AA50-22AFC1C26918}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FE9BE527-988E-4129-AA50-22AFC1C26918}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FE9BE527-988E-4129-AA50-22AFC1C26918}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FE9BE527-988E-4129-AA50-22AFC1C26918}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
